+++ app/src/main/java/com/picketing/www/business/service/show/SeatGradeService.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.picketing.www.business.domain.show.SeatGrade;
 import com.picketing.www.business.domain.show.SeatGradeFactory;
-import com.picketing.www.persistence.repository.show.ShowRepository;
+import com.picketing.www.persistence.repository.show.SeatGradeRepository;
 import com.picketing.www.persistence.table.SeatGradePersist;
 
 import lombok.RequiredArgsConstructor;
@@ -19,12 +19,11 @@ public class SeatGradeService {
 
 	private final SeatGradeFactory seatGradeFactory;
 
-	private final ShowRepository showRepository;
+	private final SeatGradeRepository seatGradeRepository;
 
 	@Transactional(readOnly = true)
 	public List<SeatGrade> gets(Long showId, Long scheduleId) {
-		// jpa 야 빨리 와라...
-		List<SeatGradePersist> seatGradePersists = showRepository.findSeatGradeByShowIdAndTime(showId, scheduleId);
+		List<SeatGradePersist> seatGradePersists = seatGradeRepository.findSeatGradeByShowIdAndTime(showId, scheduleId);
 		return seatGradePersists.stream()
 			.map(seatGradeFactory::create)
 			.toList();
