+++ app/src/main/java/com/picketing/www/persistence/repository/show/SeatGradeRepository.java
@@ -1,0 +1,9 @@
+package com.picketing.www.persistence.repository.show;
+
+import java.util.List;
+
+import com.picketing.www.persistence.table.SeatGradePersist;
+
+public interface SeatGradeRepository {
+	List<SeatGradePersist> findSeatGradeByShowIdAndTime(Long showId, Long scheduleId);
+}
